move CodeOwnerMetricMiddleware
The CodeOwnerMetricMiddleware is being moved from
edx-platform to edx-django-utils so it can be used
in multiple IDAs. The justification for this move
is in the ADR included in this commit.

This PR also contains minimal changes, like updating
paths and comments as required.

The original implementation was here:
https://github.com/edx/edx-platform/blob/7e56d89bcccf2082a9c74d91fb44b15046f00b45/lms/djangoapps/monitoring/middleware.py#L13

The original implementation will soon be removed and
replaced with this new implementation.

ARCHBOM-1356