Make heading characters consistent in docs.
The CHANGELOG and README rst files get combined into one file by
setup.py as the long description of the project on PyPI.  Howevere they
used the characters below headings inconsistently from eachother.

When we added a new level of heading in the README, it conflicted with
the ones used in the changelog.  I couldn't easily get them to match by
just updating the README.rst so I'm updating both the CHANGELOG.rst and
README.rst files to make them consistent.