feat: Add utility for logging sensitive information using encryption
This is copied from edx-platform, and was introduced there in
<https://github.com/openedx/edx-platform/pull/29682>. Bringing it out into
this library will allow it to be used across IDAs, and in particular in
the cookie monitoring middleware.

ref: ARCHBOM-2082

Introduces `log_sensitive` module for public-key encryption of sensitive
debug information in log messages, including CLI commands for generating
keys and decrypting log output.

Also:

- Adds `PyNaCl` to base requirements for encryption tools