build: Demo of no-setup Makefile (standardized virtualenv handling)
This allows a developer to run `make test` or other targets without having
to manually set up a virtualenv or install requirements. The virtualenv is
set up for them as a hidden subdirectory and requirements are automatically
installed as needed.

The goal here is larger: The new base.mk would be pulled in via git subtree
from some centralized repository. Over time, this standardized base
Makefile would grow and the Makefiles in regular repositories would shrink.
By centralizing Makefile logic, we would gain the following:

- Standardized names for make targets. Currently there is no universal way
  to install Python dependencies, to run all tests and quality checks, or
  to build docs. Developers have to learn what is used in each repo. We
  have standards for some of this, but they're not always followed. By
  composing Makefiles, we can ensure new repos always follow the standard,
  even as the standard changes.
- Faster onboarding of developers who have not yet set up their own
  personalized virtualenv toolchains. Experienced Python developers already
  have these tools set up, but newer developers stumble on virtualenv
  creation.
- A way to update repos to the latest Makefile improvements without having
  to manually update several hundreds of repos by hand. Automating the
  update of a git subtree reference still requires making and merging PRs
  but it does not require manual review of each PR, let alone manual
  PR creation.