feat: Basic OpenTelemetry and Datadog support
Switch to pluggable design so that some monitoring functions can
report to something other than New Relic. Still defaults to just New
Relic, but new setting allows adding OpenTelemetry or Datadog, or
removing New Relic.

Initialization and configuration of OpenTelemetry is left as an
exercise to the deployer, but
https://github.com/mitodl/open-edx-plugins/tree/main/src/ol_openedx_otel_monitoring/
would be a likely candidate.

Part of https://github.com/openedx/edx-django-utils/issues/389